CSHARP-3341: Unified test runner cleaning up.
