+++ tests/MongoDB.Driver.Tests/Specifications/unified-test-format/UnifiedTestFormatNegativeTestRunner.cs
@@ -18,13 +18,14 @@ using System.Collections.Generic;
 using FluentAssertions;
 using MongoDB.Bson;
 using MongoDB.Bson.TestHelpers.JsonDrivenTests;
+using MongoDB.Driver.Tests.UnifiedTestOperations;
 using Xunit;
 
 namespace MongoDB.Driver.Tests.Specifications.unified_test_format
 {
     public sealed class UnifiedTestFormatNegativeTestRunner : IDisposable
     {
-        private readonly UnifiedTestFormatTestRunner _unifiedTestFormatTestRunner = new UnifiedTestFormatTestRunner();
+        private readonly UnifiedTestFormatProcessor _unifiedTestFormatTestRunner = new UnifiedTestFormatProcessor();
 
         public void Dispose()
         {
